Update publish.yml
According to https://developercommunity.visualstudio.com/content/problem/758627/unable-to-publish-a-python-package-from-a-pipeline.html it might be that I have to set the project name in the authentication task.

However, we tried this before to no avail.  The link used the $(PROJECT_NAME) convention but I'm not sure that's not just an ad-hoc comment convention.  It doesn't seem to be an actual pipeline env variable to use.

So I'm going to try System.TeamProjectId first, followed by System.TeamProject if the id is inappropriate.