Build for 3.10
It's possible we can build for 3.10 without too much effort, so I'm going to try to do that.  I can't create a build for just 3.10 on the 1.0.0 spot because our release process doesn't make that super easy, so I'm going to increment to 1.0.1 without actually making any changes.  If I get a build working I'll see about including a type stub in our package as well for 1.0.1 so people can rely on that too