Build with New Deps
There's a [CVE](https://nvd.nist.gov/vuln/detail/CVE-2020-26235) out for the `time` crate, which `chrono` depends on. Our last build was so long ago that it was built using a version of `time` that had this issue.  

In theory, merely making a new build will rectify this.

In theory.