Minor fixes in packaging to make it compatible with maturin 1.4.0 (#39)
* Build package with maturin 1.4.0

* Loosen maturin version requirement

* Remove pyo3 from install instructions in README

* Remove empty package.metadata.maturin section

* Fix the builds

* Update action versions

* Lock maturin to major version

* Bump version of python package to 1.2.1

* Add building src-dist in ubuntu build

* Update build.yml

Writing out the sha256 checksum for the sdist in `dist/`

* Update README.md

We should probably support 3.12 as well, since it's a final release. It *should* work, since we're specifically only targeting the broader python 3 abi with this. I should write some simple unit tests that run across the full matrix of python versions. That's a fix for me to do, though.

* Update pyproject.toml

3.12 support! We think!

---------

Co-authored-by: Dax Pryce <daxpryce@microsoft.com>